requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
distlib==0.3.7
dnspython==2.7.0
ecdsa==0.18.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.103.1
filelock==3.12.4
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.0
httpx==0.25.1
idna==3.4
inflect==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
mysql-connector-python==8.3.0
packaging==23.1
pipenv==2023.9.8
platformdirs==3.10.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
virtualenv==20.24.5
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
